refactor: convert Bell.php to single-file Livewire 4 component, remove separate class file
